Updated compose ui files/ cleaned inherited classes/moved parent class to folder
