+++ app/src/main/java/app/plugbrain/android/ui/challenges/compose/NumericalChallengeScreen.kt
@@ -16,7 +16,7 @@ import app.plugbrain.android.BuildConfig
 import app.plugbrain.android.challenges.MissingChallenge
 import app.plugbrain.android.challenges.NumericalChallenge
 import app.plugbrain.android.challenges.TwoOperandsChallenge
-import app.plugbrain.android.challenges.SquareChallenge
+import app.plugbrain.android.challenges.square.SquareChallenge
 import app.plugbrain.android.challenges.addition.AdditionTwoDigitsCarryFreeChallenge
 import app.plugbrain.android.ui.isPortrait
